prove typing_through_subst_ee
Note that we need to add *value u* as a hypothesis, otherwise it can't
be proved. This is because if u is application(t1 t2), then t1 and t2
are typed separately, typing of t1 and t2 may depend on variables that
are not in (closed_env E). When T and U are both closed types, the first
typing holds, but the second not.

Lemma typing_through_subst_ee : forall U E F x T e u,
  value u ->
  typing (E & x ~: U & F) e T ->               (* 1 *)
  typing E u U ->
  typing (E & F) (subst_ee x u e) T.           (* 2 *)